requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
Django==3.1.5
Flask==1.1.2
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
psycopg2==2.8.6
pyee==7.0.4
python-dotenv==0.15.0
pytz==2020.5
slackclient==2.9.3
slackeventsapi==2.2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.3
